Protect BLOB integrity using hash instead of HMAC
We don't expect a forgery of the entire BLOB. That would require an
attacker to forge each block. So let's use a hash instead of HMAC. This
allows to transmit a BLOB without having a common key at the moment of
BLOB creation and last block reception.

At the same time, harden BLOB header by requiring 4 zero bytes.

All of these should be enough to protect BLOB integrity in case of
unexpected errors or block misauthentication.